Simpler definition and new proof for CoreDefunc.
The new proof operates on the category of binary
relations over sets, and makes use of the fact
that strong bisimulation can be understood as
isomorphicity (in here, with isomorphism d) in that
category.

I plan to use virtually the same argument with
CoreRefunc; maybe can factor Lemma 4.2.2 out such
that it only depends on some isomorphism, not d specifically. Then, the following would suffice
for CoreRefunc: give an r^{-1} and a lemma for r,
like Lemma 4.2.1 is for d.